NETOBSERV-2848: Split test files and switch to demoLoki
